requirements.txt
mcp
starlette
uvicorn
ollama
chromadb
Pillow
watchdog
pymupdf
openpyxl
python-docx
python-pptx
